requirements.txt
agency-swarm==0.4.2
python-dotenv>=0.19.0
pydantic>=2.0.0 